backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
geopandas
shapely
fiona
aiofiles
python-multipart
requests
pycurl
httpx
playwright
aiohttp
splinter
selenium
pyinstaller
geojson
bs4
pyproj
uuid
pdfplumber
pytesseract 
pdf2image
deskewed_page2
rapidfuzz
paddlepaddle==2.5.2
paddleocr
matplotlib
requests
paddleocr[structure]